[softek-config] Created directory for backups
Moved files that are backups or screenshots or what have you to their
own folder. I'm still in the process of figuring out how I want to
organize this, but my vague plan is to put exports of settings for
various programs into this repo, including stuff for Windows products
(though it pains me to have to do it).